Migrate some tests, add more
